// frontend/src/shared/lib/api/model/plateGroupDetailResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Cellar
 * OpenAPI spec version: 0.1.0
 */
import type { PlateGroupResponse } from './plateGroupResponse';
import type { PlateGroupDetailResponsePlateFormat } from './plateGroupDetailResponsePlateFormat';
import type { GroupRefResponse } from './groupRefResponse';

export interface PlateGroupDetailResponse {
  group: PlateGroupResponse;
  plate_count: number;
  subtree_plate_count: number;
  plate_format?: PlateGroupDetailResponsePlateFormat;
  ancestors: GroupRefResponse[];
  children: GroupRefResponse[];
}